修改首页查询无法显示正常数据

+++ phoibe-doc-cms-domain/src/main/java/me/phoibe/doc/cms/domain/po/PhoibeUserPostil.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class PhoibeUserPostil {
     private int id;
     private long userId;
-    private int docId;
+    private long docId;
     private String docName;
     private String docPath;
     private Date createTime;
@@ -37,7 +37,7 @@ public class PhoibeUserPostil {
         this.userId = userId;
     }
 
-    public int getDocId() {
+    public long getDocId() {
         return docId;
     }
